Use new #languages= method in data import script
Currently EthnicChurch#languages= takes an array of strings (for
existing and non-existing records). This is perhaps not the best
implementation as the default active record method takes Language
objects. In the future, this will probably be re-implemented but for now
this commit keeps the import script in sync with the current
implementation.